staff api show list of clients

+++ db/migrate/20190916091718_add_fullname_phone_to_clients.rb
@@ -1,6 +1,6 @@
 class AddFullnamePhoneToClients < ActiveRecord::Migration[6.0]
   def change
     add_column :clients, :fullname, :string, null: false
-    add_column :clients, :phone, :integer, null: false
+    add_column :clients, :phone, :bigint, null: false
   end
 end
